Errors in tests stops execution.
In 1.2.0 the behavior was to throw an exception on error. This resulted in a stacktrace pointing to the plugin (confusing behavior). I removed it for version 1.2.1 but forgot that it will cause the runner to continue after a failed test.

This change makes it trigger a system.exit(1) upon failure.