Handling different exceptions based on type of exception.
A parsing error (stuff like: x is not defined) will now print a complete javascript stacktrace.

If there for some reason is a java exception thrown (test that a using the rhino java environment for example java.io.File or similar) this error will also show a complete stack trace.

If there is a regular javascript exception only the error message will be shown. I could not find any way to actually print a complete stacktrace to that error due to a rhino bug. So if you do any stuff like this in you javascript:

    throw "an error";

This will not have a stacktrace.